fix for detail and impl
