+++ include/aquarius.hpp
@@ -16,7 +16,7 @@
 #include <aquarius/executor/tcp_client.hpp>
 #include <aquarius/executor/tcp_server.hpp>
 #include <aquarius/flex_buffer.hpp>
-#include <aquarius/detail/impl/context.hpp>
+#include <aquarius/flow_context.hpp>
 #include <aquarius/io_context.hpp>
 #include <aquarius/logger.hpp>
 #include <aquarius/noncopyable.hpp>
