--- include/aquarius/detail/impl/io_object_impl.hpp
@@ -1,13 +1,0 @@
-#pragma once
-#include <boost/asio/detail/io_object_impl.hpp>
-
-namespace aquarius
-{
-	namespace detail
-	{
-		namespace impl
-		{
-			using boost::asio::detail::io_object_impl;
-		}
-	} // namespace detail
-} // namespace aquarius
