+++ include/aquarius/detail/session_service.hpp
@@ -1,11 +1,11 @@
 #pragma once
 #include <aquarius/detail/config.hpp>
-#include <aquarius/detail/impl/session_service.hpp>
+#include <aquarius/detail/basic_session_service.hpp>
 
 namespace aquarius
 {
 	namespace detail
 	{
-		using session_service = impl::session_service;
+		using session_service = session_service;
 	}
 } // namespace aquarius
