+++ include/aquarius/flow_context.hpp
@@ -2,11 +2,11 @@
 #include <aquarius/detail/config.hpp>
 // #include <aquarius/executor/tcp_client_session.hpp>
 // #include <aquarius/executor/tcp_server_session.hpp>
-#include <aquarius/detail/impl/handler.hpp>
+#include <aquarius/detail/handler.hpp>
 
 #define AQUARIUS_CONTEXT_BY(__session, method, __rpc)                                                                  \
 	class method;                                                                                                      \
-	[[maybe_unused]] static aquarius::detail::impl::auto_handler_register<__session, __rpc, method>                    \
+	[[maybe_unused]] static aquarius::detail::auto_handler_register<__session, __rpc, method>                    \
 		__auto_register_##method(__rpc::id);                                                                           \
 	AQUARIUS_POSITIVE_HANDLER(__session, method, __rpc::request, __rpc::response)
 
@@ -16,6 +16,6 @@
 
 #define AQUARIUS_RESPONSE(method, __rpc)                                                                               \
 	class method;                                                                                                      \
-	[[maybe_unused]] static aquarius::detail::impl::auto_handler_register<aquarius::tcp_client_session, __rpc, method> \
+	[[maybe_unused]] static aquarius::detail::auto_handler_register<aquarius::tcp_client_session, __rpc, method> \
 		__auto_register_##method(__rpc::id);                                                                           \
 	AQUARIUS_HANDLER(aquarius::tcp_client_session, method, __rpc::response)
